requirements.txt
Django
django-ckeditor
django-imagekit
pytz
django-disqus
django-crispy-forms